fix( sessionStorage ): refacto de tous les endroit où le token était cherché dans localStorage à la place de sessionStorage, ce qui entrainait un bug ( forcément ) ; need to fix : nettoyage des console.log pour le debuggage !

+++ FrontEnd/assets/js/models/deleteModel.js
@@ -1,6 +1,11 @@
 const API_BASE_URL = 'http://localhost:5678/api';
 
 export async function deleteWorkById(id, token) {
+  if (!id) {
+    console.error('ID non défini lors de la tentative de suppression');
+    throw new Error('ID manquant');
+  }
+
   const response = await fetch(`${API_BASE_URL}/works/${id}`, {
     method: 'DELETE',
     headers: {
@@ -9,8 +14,10 @@ export async function deleteWorkById(id, token) {
   });
 
   if (!response.ok) {
-    throw new Error('Erreur lors de la suppression');
+    const errorData = await response.json();  // Récupérer les données d'erreur
+    console.error('Erreur lors de la suppression:', errorData);
+    throw new Error(errorData.message || 'Erreur lors de la suppression');
   }
-
+  
   return true;
 }
